php warnings
This converts the warnings for php security and php not up to date from the horrible bs-collapse to use the native details/summary which is used elsewhere.

This has the bonus that it is much clearer now that you can click to expand and you dont have to use inline css to change the pointer.

I left the pep eol message as is because that is justy displayed without a disclosure element.

Signed-off-by: brian teeman <brian@teeman.net>