RFC Column Link Title
As discussed before. The way I see this there are two options.

### Option 1
_as shown here_

The helper already supports a second string which is used for the title. That's not perfect for a screenreader as it sees it as a description which will only be announced in verbose mode. But its better than the hash and also means you get a title tooltip with some meaning

### Option 2
This would be my preferred option and the one which matches your original concept of having a separate string for both sets of users. Simply add an extra param upstream to the awf helper which will produce an aria-label on the link. This would be the _best_ for accessibility

Signed-off-by: brian teeman <brian@teeman.net>