navbar forced dark-mode? (#454)
I couldn't understand why the dropdown menu for automations always had a dark background. It's because it has a hardcoded `data-bs-theme=dark`

For me it is confusing, and a misuse of that attribute to use it this way.

This PR removes the data-bs-theme and adjusts the class of the title. The automations dropdown is now styled the same as the other dropdown menus.

Signed-off-by: brian teeman <brian@teeman.net>